chore: Cleanup PR after review
